# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI Data Analyst API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Dataset {
  id: string;
  name: string;
  filename: string;
  rowCount: number;
  columnCount: number;
  uploadedAt: string;
  fileSize: number;
}

export type DatasetDetailPreviewRowsItem = { [key: string]: unknown };

export interface ColumnInfo {
  name: string;
  dtype: string;
  nullCount: number;
  uniqueCount: number;
  /** @nullable */
  min?: string | null;
  /** @nullable */
  max?: string | null;
  /** @nullable */
  mean?: number | null;
  /** @nullable */
  std?: number | null;
}

export interface DatasetDetail {
  id: string;
  name: string;
  filename: string;
  rowCount: number;
  columnCount: number;
  uploadedAt: string;
  fileSize: number;
  columns: ColumnInfo[];
  previewRows?: DatasetDetailPreviewRowsItem[];
}

export interface DatasetOverview {
  rowCount: number;
  columnCount: number;
  missingCells: number;
  missingPercent?: number;
  duplicateRows: number;
  numericColumns: number;
  categoricalColumns: number;
  columns: ColumnInfo[];
}

export type ChartConfigDataItem = { [key: string]: unknown };

export interface ChartConfig {
  id: string;
  title: string;
  chartType: string;
  xAxis: string;
  yAxis: string;
  /** @nullable */
  description?: string | null;
  data: ChartConfigDataItem[];
}

export interface KpiMetric {
  id: string;
  label: string;
  value: string;
  column: string;
  /** @nullable */
  change?: number | null;
  /** @nullable */
  trend?: string | null;
}

export interface InsightsResult {
  insights: string;
  generatedAt: string;
  keyFindings?: string[];
}

export interface ChatMessageInput {
  message: string;
}

export interface ChatMessage {
  id: string;
  role: string;
  content: string;
  createdAt: string;
}

export interface ChatResponse {
  id: string;
  role: string;
  content: string;
  createdAt: string;
}

export interface ReportInput {
  title?: string;
  includeCharts?: boolean;
  includeInsights?: boolean;
  includeKpis?: boolean;
}

export interface ReportResult {
  id: string;
  datasetId: string;
  title: string;
  createdAt: string;
  downloadUrl: string;
}

export interface ReportMeta {
  id: string;
  datasetId: string;
  datasetName: string;
  title: string;
  createdAt: string;
  downloadUrl: string;
}

export interface DeleteResult {
  success: boolean;
  /** @nullable */
  message?: string | null;
}