# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: AI Data Analyst API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /datasets/{id}/overview:
    get:
      operationId: getDatasetOverview
      tags: [analysis]
      summary: Get statistical overview of the dataset
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Dataset overview
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DatasetOverview"
  /datasets/{id}/charts:
    get:
      operationId: getDatasetCharts
      tags: [analysis]
      summary: Get auto-generated charts for the dataset
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of chart configs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChartConfig"
  /datasets/{id}/kpis:
    get:
      operationId: getDatasetKpis
      tags: [analysis]
      summary: Get KPI metrics for the dataset
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: KPI metrics
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/KpiMetric"
  /datasets/{id}/insights:
    get:
      operationId: getDatasetInsights
      tags: [analysis]
      summary: Get AI-generated insights for the dataset
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: AI insights
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InsightsResult"
  /datasets/{id}/chat:
    post:
      operationId: chatWithDataset
      tags: [chat]
      summary: Ask a question about the dataset
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatMessageInput"
      responses:
        "200":
          description: AI response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatResponse"
  /datasets/{id}/chat/history:
    get:
      operationId: getChatHistory
      tags: [chat]
      summary: Get chat history for a dataset
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Chat history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChatMessage"
    delete:
      operationId: clearChatHistory
      tags: [chat]
      summary: Clear chat history for a dataset
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Cleared
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResult"
...
components:
  schemas:
...
    ColumnInfo:
      type: object
      required: [name, dtype, nullCount, uniqueCount]
      properties:
        name:
          type: string
        dtype:
          type: string
        nullCount:
          type: integer
        uniqueCount:
          type: integer
        min:
          type: ["string", "null"]
        max:
          type: ["string", "null"]
        mean:
          type: ["number", "null"]
        std:
          type: ["number", "null"]
    DatasetOverview:
      type: object
      required: [rowCount, columnCount, missingCells, duplicateRows, numericColumns, categoricalColumns, columns]
      properties:
        rowCount:
          type: integer
        columnCount:
          type: integer
        missingCells:
          type: integer
        missingPercent:
          type: number
        duplicateRows:
          type: integer
        numericColumns:
          type: integer
        categoricalColumns:
          type: integer
        columns:
          type: array
          items:
            $ref: "#/components/schemas/ColumnInfo"
    ChartConfig:
      type: object
      required: [id, title, chartType, xAxis, yAxis, data]
      properties:
        id:
          type: string
        title:
          type: string
        chartType:
          type: string
        xAxis:
          type: string
        yAxis:
          type: string
        description:
          type: ["string", "null"]
        data:
          type: array
          items:
            type: object
    KpiMetric:
      type: object
      required: [id, label, value, column]
      properties:
        id:
          type: string
        label:
          type: string
        value:
          type: string
        column:
          type: string
        change:
          type: ["number", "null"]
        trend:
          type: ["string", "null"]
    InsightsResult:
      type: object
      required: [insights, generatedAt]
      properties:
        insights:
          type: string
        generatedAt:
          type: string
        keyFindings:
          type: array
          items:
            type: string
    ChatMessageInput:
      type: object
      required: [message]
      properties:
        message:
          type: string
    ChatMessage:
      type: object
      required: [id, role, content, createdAt]
      properties:
        id:
          type: string
        role:
          type: string
        content:
          type: string
        createdAt:
          type: string
    ChatResponse:
      type: object
      required: [id, role, content, createdAt]
      properties:
        id:
          type: string
        role:
          type: string
        content:
          type: string
        createdAt:
          type: string
...
    DeleteResult:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
        message:
          type: ["string", "null"]