# lib/api-zod/src/generated/types/columnInfo.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI Data Analyst API
 * OpenAPI spec version: 0.1.0
 */

export interface ColumnInfo {
  name: string;
  dtype: string;
  nullCount: number;
  uniqueCount: number;
  /** @nullable */
  min?: string | null;
  /** @nullable */
  max?: string | null;
  /** @nullable */
  mean?: number | null;
  /** @nullable */
  std?: number | null;
}