# lib/api-zod/src/generated/types/datasetOverview.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI Data Analyst API
 * OpenAPI spec version: 0.1.0
 */
import type { ColumnInfo } from './columnInfo';

export interface DatasetOverview {
  rowCount: number;
  columnCount: number;
  missingCells: number;
  missingPercent?: number;
  duplicateRows: number;
  numericColumns: number;
  categoricalColumns: number;
  columns: ColumnInfo[];
}