# lib/api-zod/src/generated/types/kpiMetric.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI Data Analyst API
 * OpenAPI spec version: 0.1.0
 */

export interface KpiMetric {
  id: string;
  label: string;
  value: string;
  column: string;
  /** @nullable */
  change?: number | null;
  /** @nullable */
  trend?: string | null;
}